+++ src/main/java/com/kodilla/ecommercee/controller/cart/DBCartService.java
@@ -1,0 +1,7 @@
+package com.kodilla.ecommercee.controller.cart;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class DBCartService {
+}
