+++ src/main/java/com/kodilla/ecommercee/domain/cart/CartDto.java
@@ -1,3 +1,4 @@
+
 package com.kodilla.ecommercee.domain.cart;
 
 import lombok.AllArgsConstructor;
@@ -7,11 +8,9 @@ import lombok.NoArgsConstructor;
 import java.math.BigDecimal;
 
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
-public class ProductDtoToCart {
-
+@NoArgsConstructor
+public class CartDto {
     private Long id;
-    private String description;
-    private BigDecimal price;
+    private BigDecimal cost;
 }
